refactoring to remove datafusion subpackage
