+++ native/core/src/execution/expressions/mod.rs
@@ -21,4 +21,4 @@ pub mod bloom_filter_agg;
 pub mod bloom_filter_might_contain;
 pub mod subquery;
 
-pub use datafusion_comet_spark_expr::{EvalMode, SparkError};
+pub use datafusion_comet_spark_expr::EvalMode;
