+++ native/core/src/execution/mod.rs
@@ -16,13 +16,16 @@
 // under the License.
 
 //! PoC of vectorization execution through JNI to Rust.
-pub mod datafusion;
+pub(crate) mod expressions;
 pub mod jni_api;
 mod metrics;
 pub mod operators;
+pub(crate) mod planner;
 pub mod serde;
 pub mod shuffle;
 pub(crate) mod sort;
+pub(crate) mod spark_plan;
+pub(crate) mod util;
 pub use datafusion_comet_spark_expr::timezone;
 pub(crate) mod utils;
