+++ native/core/src/execution/shuffle/mod.rs
@@ -18,3 +18,5 @@
 mod list;
 mod map;
 pub mod row;
+mod shuffle_writer;
+pub use shuffle_writer::ShuffleWriterExec;
